Started Custodial Update
The custodial limit requires careful handling of the degenerate masses to avoid divergences. I started modifying the functions.
-matchResToUVScanCard[] has a new option called "DegenerateMasses" that will allow the treatment of those particular cases. The default is "False".
When that option is "True", it will call to a new "massHandlerCusto" function instead of the usual "massHandler"
The "massHandlerCusto" is at the moment just a copy of "massHandler", lacks proper implementation.